pyproject.toml
[tool.poetry]
name = "argo-workflow-tools"
version = "0.8.0"
description = "A suite of tools to ease ML pipeline development with Argo Workflows"
authors = ["Diagnostic Robotics"]
license = "Apache 2.0"
repository = "https://github.com/DiagnosticRobotics/argo-workflow-tools"
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.7"
PyYAML = "^6.0"
pydantic = "^1.7.4"
contextvars = "^2.4"
shortuuid = "^1.0.8"
urllib3 = "^1.26.7"
certifi = "^2021.10.8"
requests = "^2.26.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
black = "^21.11b"
isort = "^5.7.0"
deepdiff = "^5.2.3"
flake8 = "^4.0.1"
flake8-tidy-imports = "^4.5.0"
pycodestyle = "^2.8.0"
pyflakes = "^2.4.0"
flake8-pytest-style = "^1.5.1"
flake8-quotes = "^3.3.1"
pydocstyle = "^6.1.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.pydocstyle]
inherit = false
convention = "numpy"